library-app/requirements.txt
alembic==1.9.2
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.3
black==24.4.2
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastapi-pagination==0.12.24
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.3
packaging==24.0
paginate==0.5.6
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.4
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==1.4.6
sqlalchemy-utils==0.41.2
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
unicode-tr==0.6.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
celery==5.4.0
fastapi-mail==1.4.1
asgiref==3.8.1
openpyxl==3.1.2